split test_storage.py and test_misc.py
