+++ tests/test_aws_utils.py
@@ -4,7 +4,6 @@ from unittest.mock import patch
 from tests.conftest import TEST_USER_ID
 from gen3workflow.aws import aws_utils, clients, bucket
 from gen3workflow.config import config
-from tests.test_misc import mock_aws_services
 
 
 def test_create_role_for_bucket_access_creates_role_when_missing(mock_aws_services):
